fix(build): avoid eigenlayer deps in tangle-only builds
